fix STUPID assertion error

+++ pi/serial_test.py
@@ -6,7 +6,7 @@ from tqdm import tqdm
 # from simple_command import compute_motor_values
 # from new_closed_loop import compute_motor_values
 from visual_controller import compute_motor_values, cam
-from numpy import int64, float64
+from numpy import int32, float64
 
 
 debug_mode = False
@@ -25,8 +25,8 @@ with Serial(port=ports[0], baudrate=115200) as ser:
     def write_motors(left_motor, right_motor):
         global last_write, left_motor_prev, right_motor_prev
 
-        assert type(left_motor) in [int, float, int64, float64], "motor input should be of type int, not {}".format(type(left_motor))
-        assert type(right_motor) in [int, float, int64, float64], "motor input should be of type int, not {}".format(type(right_motor))
+        assert type(left_motor) in [int, float, int32, float64], "motor input should be of type int, not {}".format(type(left_motor))
+        assert type(right_motor) in [int, float, int32, float64], "motor input should be of type int, not {}".format(type(right_motor))
         left_motor, right_motor = int(left_motor), int(right_motor)
 
         last_write = time()
